# backend/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API for the Blind Dating App 'Title in Progress'",
		"version": "1.0.0",
		"title": "Blind Date",
		"contact": {
			"email": "anne.huesges@hpe.com"
		}
	},
	"host": "www.example.de",
	"tags": [
		{
			"name": "chatMsg",
			"description": "Messages between our users"
		},
		{
			"name": "user",
			"description": "Operations concerning users"
		},
		{
			"name": "profile",
			"description": "Interactions with the user profile"
		}
	],
	"schemes": ["https", "http"],
	"paths": {
		"/profile": {
			"post": {
				"tags": ["profile"],
				"summary": "Add new profile information",
				"description": "",
				"operationId": "addProfile",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"description": "Profile Object",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Profile"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Profile added successfully"
					},
					"405": {
						"description": "Invalid input"
					}
				}
			}
		},
		"/profile/{id}": {
			"get": {
				"tags": ["profile"],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "The user id of the user, whom the profile belongs to",
						"required": true,
						"type": "integer"
					}
				],
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "Profile added successfully",
						"schema": {
							"$ref": "#/definitions/Profile"
						}
					},
					"405": {
						"description": "Invalid input"
					}
				}
			}
		},
		"/user": {
			"post": {
				"tags": ["user"],
				"summary": "Add a new User",
				"description": "",
				"operationId": "addUser",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"description": "User Object",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"405": {
						"description": "Invalid input"
					},
					"409": {
						"description": "Username already taken"
					},
					"422": {
						"description": "Invalid username supplied"
					}
				},
				"security": [
					{
						"petstore_auth": ["write:pets", "read:pets"]
					}
				]
			}
		},
		"/user/login": {
			"get": {
				"tags": ["user"],
				"summary": "Logs user into the system",
				"description": "",
				"operationId": "loginUser",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "username",
						"in": "query",
						"description": "The user name for login",
						"required": true,
						"type": "string"
					},
					{
						"name": "password",
						"in": "query",
						"description": "The password for login in clear text",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"type": "string"
						},
						"headers": {
							"X-Rate-Limit": {
								"type": "integer",
								"format": "int32",
								"description": "calls per hour allowed by the user"
							},
							"X-Expires-After": {
								"type": "string",
								"format": "date-time",
								"description": "date in UTC when token expires"
							}
						}
					},
					"400": {
						"description": "Invalid username/password supplied"
					}
				}
			}
		},
		"/user/logout": {
			"get": {
				"tags": ["user"],
				"summary": "Logs out current logged in user session",
				"description": "",
				"operationId": "logoutUser",
				"produces": ["application/json"],
				"parameters": [],
				"responses": {
					"default": {
						"description": "successful operation"
					}
				}
			}
		},
		"/user/{username}": {
			"get": {
				"tags": ["user"],
				"summary": "Get user by user name",
				"description": "",
				"operationId": "getUserByName",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "username",
						"in": "path",
						"description": "The name that needs to be fetched. Use user1 for testing. ",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/User"
						}
					},
					"400": {
						"description": "Invalid username supplied"
					},
					"404": {
						"description": "User not found"
					}
				}
			},
			"put": {
				"tags": ["user"],
				"summary": "Updated user",
				"description": "This can only be done by the logged in user.",
				"operationId": "updateUser",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "username",
						"in": "path",
						"description": "name that need to be updated",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Updated user object",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"400": {
						"description": "Invalid user supplied"
					},
					"404": {
						"description": "User not found"
					}
				}
			},
			"delete": {
				"tags": ["user"],
				"summary": "Delete user",
				"description": "This can only be done by the logged in user.",
				"operationId": "deleteUser",
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "username",
						"in": "path",
						"description": "The username of the account to be deleted",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"400": {
						"description": "Invalid username supplied"
					},
					"404": {
						"description": "User not found"
					}
				}
			}
		}
	},
	"securityDefinitions": {
		"petstore_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify pets in your account",
				"read:pets": "read your pets"
			}
		},
		"api_key": {
			"type": "apiKey",
			"name": "api_key",
			"in": "header"
		}
	},
	"definitions": {
		"Profile": {
			"type": "object",
			"properties": {
				"user": {
					"$ref": "#/definitions/User"
				},
				"firstName": {
					"type": "string",
					"example": "Max"
				},
				"lastName": {
					"type": "string",
					"example": "Mustermann"
				},
				"age": {
					"type": "integer",
					"format": "int64",
					"example": "25"
				},
				"hobbies": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"example": ["tennis", "football"]
				},
				"gender": {
					"type": "string",
					"example": "male"
				},
				"preferences": {
					"type": "object",
					"properties": {
						"genderPref": {
							"type": "string",
							"example": "male"
						},
						"ageRange": {
							"type": "object",
							"properties": {
								"minAge": {
									"type": "integer",
									"format": "int64",
									"example": "25"
								},
								"maxAge": {
									"type": "integer",
									"format": "int64",
									"example": "30"
								}
							}
						}
					}
				},
				"images": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
		"User": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64",
					"example": "10"
				},
				"email": {
					"type": "string",
					"example": "max.muster@mail.de"
				},
				"username": {
					"type": "string",
					"example": "Max Mustermann"
				},
				"password": {
					"type": "string",
					"example": "PlzN0t1234"
				}
			}
		},
		"ApiResponse": {
			"type": "object",
			"properties": {
				"code": {
					"type": "integer",
					"format": "int32"
				},
				"type": {
					"type": "string"
				},
				"message": {
					"type": "string"
				}
			}
		}
	}
}